refactor: folders and files hierarchy
