+++ src/app/helpers/helpFunction.ts
