+++ src/index.ts
@@ -1,5 +1,5 @@
 import './style.scss';
-import { gallery } from './app/mainGallery';
-import { filter } from './app/filters/filters';
+import { gallery } from './app/components/mainGallery/mainGallery';
+import { filter } from './app/components/filters/filters';
 gallery.createGallery();
 filter.createAll();
